kainexa-core/requirements.txt
# requirements.txt (alternative to poetry)
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.0
sqlalchemy==2.0.0
asyncpg==0.29.0
alembic==1.12.0
redis==5.0.0
transformers==4.56.2
torch==2.4.1+cu121
huggingface-hub==0.35.1
konlpy==0.6.0
kiwipiepy==0.17.0
aiohttp==3.9.0
httpx==0.25.0
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
distro==1.9.0 
jiter==0.11.0 
langchain-openai==0.3.33 
openai==1.109.0
# LangChain 0.3 계열(모듈 분리됨)
langchain>=0.3.27
langchain-community>=0.3.29
langchain-text-splitters>=0.3.11
langchain-huggingface==0.3.1
# 벡터스토어: Chroma (로컬 퍼시스턴스, CUDA 무관)
chromadb==1.1.0
# 임베딩(HF): sentence-transformers 기반
sentence-transformers==5.1.1

# 문서 로더용 의존성
python-docx==1.2.0
pypdf==6.1.0
beautifulsoup4==4.13.5
html2text==2025.4.15
soupsieve==2.8

# 토크나이저
tiktoken==0.11.0

PyJWT[crypto]>=2.8.0 # cryptography 포함
qdrant-client>=1.14,<2.0
python-multipart>=0.0.9
aiofiles>=23.1.0  #비동기 파일 저장

accelerate>=0.34.0
bitsandbytes>=0.41.0

# 추가 package 등록